DihedralParametrization

Documentation for DihedralParametrization.

DihedralParametrization.atomcoordinates

DihedralParametrization.bondparametrization

DihedralParametrization.buildchain

DihedralParametrization.snnerf

DihedralParametrization.atomcoordinates — Method

X = atomcoordinates(bp::BondParametrization{T}, dihedrals::Vector{T}, n::AbstractVector{T}, cα::AbstractVector{T}, c::AbstractVector{T}) where T<:Real

Given a BondParametrization object bp, a vector of dihedral angles dihedrals, and the coordinates of the first three backbone atoms in the chain (n, cα, and c), compute the 3D coordinates of all atoms in the chain.

DihedralParametrization.bondparametrization — Method

bp, dihedrals = bondparametrization(chain)

Parametrize a protein chain via bond parameters. Together, bp and dihedrals are sufficient to reconstruct the 3D coordinates of the chain modulo a rigid body transformation.

dihedrals is a vector representing the rotatable dihedral angles in the chain. bp is a BondParametrization object containing all other necessary information.

DihedralParametrization.buildchain — Method

out = buildchain(reference::Chain, bp::BondParametrization, X::AbstractVector{<:SVector{3}})

Given a reference Chain object, a BondParametrization object bp, and a vector of atom 3D coordinates X, construct a new Chain object with the same sequence and atoms as reference but with the coordinates from X.

DihedralParametrization.snnerf — Method

d = snnerf(a, b, c, ℓcd, θbcd, ϕbc)

Given three points a, b, and c (as 3D vectors), the bond length ℓcd, the bond angle θbcd (in radians), and the dihedral angle ϕbc (in radians), compute the coordinates of point d using the SN-NeRF algorithm.

Reference

Parsons, Jerod, et al. "Practical conversion from torsion space to Cartesian space for in silico protein synthesis." Journal of computational chemistry 26.10 (2005): 1063-1068.